+++ R/utils.R
@@ -166,3 +166,34 @@ tags <- lapply(known_tags, function(tagname) {
     tag(tagname, contents)
   }
 })
+
+
+#' Namespaced IDs for inputs/outputs
+#'
+#' @param namespace The character vector to use for the namespace. 
+#' This can have any length, though a single element is most common. 
+#' Length 0 will cause the id to be returned without a namespace, and 
+#' length 2 will be interpreted as multiple namespaces, in increasing 
+#' order of specificity (i.e. starting with the top-level namespace).
+#' @param id The id string to be namespaced (optional).
+NS <- function (namespace, id = NULL) {
+  if (length(namespace) == 0) 
+    ns_prefix <- character(0)
+  else ns_prefix <- paste(namespace, collapse = ns.sep)
+  f <- function(id) {
+    if (length(id) == 0) 
+      return(ns_prefix)
+    if (length(ns_prefix) == 0) 
+      return(id)
+    paste(ns_prefix, id, sep = ns.sep)
+  }
+  if (missing(id)) {
+    f
+  }
+  else {
+    f(id)
+  }
+}
+
+# make R CMD check happy
+globalVariables("ns.sep")